requirements.txt
alembic==1.11.1
anyio==3.7.0
black==23.3.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.99.1
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pydantic==1.10.10
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.22.0
